Added check to redirect if already signed-in, to use server-side check as it's not client side compatible
